Extract reusable JobTable component
Created JobTable component that consolidates the job table logic from
Queue and Dashboard pages into a single reusable component.

Features:
- Configurable priority controls (queue position, move up/down/top/bottom)
- Built-in pagination with PaginationControl
- Persists rows-per-page preference to localStorage
- Consistent table structure across both pages

Updated Queue.jsx and Dashboard.jsx to use the new component:
- Queue: showPriorityControls=true for full queue management
- Dashboard: showPriorityControls=false for read-only view
